third_party/WebKit/LayoutTests/fast/events/select-event-on-input-and-textarea-expected.txt

Test select event is triggered on input and textarea Elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS select event is triggered on focus, select, setSelectionRange, setRangeText, selectionStart, selectionEnd and selectionDirection methods of input element.
PASS select event is triggered on focus, select, setSelectionRange, setRangeText, selectionStart, selectionEnd and selectionDirection methods of textarea element.
PASS successfullyParsed is true

TEST COMPLETE
 
